add power.py
